+++ src/Dfe.ManageSchoolImprovement/Models/SupportProject/ContactViewModel.cs
@@ -6,6 +6,7 @@
         public string RoleName { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
         public string Phone { get; set; } = string.Empty;
+        public string Address { get; set; } = string.Empty;
         public DateTime CreatedOn { get; set; }
         public string CreatedBy { get; set; } = string.Empty;