+++ src/Dfe.ManageSchoolImprovement/Models/SupportProject/SupportProjectViewModel.cs
@@ -1,6 +1,7 @@
 using Dfe.ManageSchoolImprovement.Application.SupportProject.Models;
 using Dfe.ManageSchoolImprovement.Domain.Entities.SupportProject;
 using Dfe.ManageSchoolImprovement.Domain.ValueObjects;
+using GovUK.Dfe.CoreLibs.Contracts.Academies.Base;
 
 namespace Dfe.ManageSchoolImprovement.Frontend.Models.SupportProject
 {
@@ -79,7 +80,6 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Models.SupportProject
         public string? AdviserFullName { get; private set; }
 
         public DateTime? IntroductoryEmailSentDate { get; set; }
-        public bool? HasShareEmailTemplateWithAdviser { get; set; }
 
         public bool? RemindAdviserToCopyRiseTeamWhenSentEmail { get; set; }
 
@@ -104,6 +104,8 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Models.SupportProject
         public string? SupportOrganisationName { get; set; }
 
         public string? SupportOrganisationIdNumber { get; set; }
+        
+        public string? SupportingOrganisationAddress { get; set; }
 
         public bool? CheckOrganisationHasCapacityAndWillingToProvideSupport { get; set; }
 
@@ -121,11 +123,10 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Models.SupportProject
 
 
         public DateTime? DateSupportingOrganisationContactDetailsAdded { get; set; }
-
         public string? SupportingOrganisationContactName { get; set; }
-
         public string? SupportingOrganisationContactEmailAddress { get; set; }
-
+        public string? SupportingOrganisationContactPhone { get; set; }
+        public string? SupportingOrganisationContactAddress { get; set; }
         public bool? IndicativeFundingBandCalculated { get; set; }
         public string? IndicativeFundingBand { get; set; }
         public bool? ImprovementPlanAndExpenditurePlanWithIndicativeFundingBandSentToSupportingOrganisationAndSchoolsResponsibleBody { get; set; }
@@ -203,7 +204,6 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Models.SupportProject
                 ResponsibleBodyResponseToTheConflictOfInterestRequestSavedInSharePoint = supportProjectDto.ResponsibleBodyResponseToTheConflictOfInterestRequestSavedInSharePoint,
                 DateConflictOfInterestDeclarationChecked = supportProjectDto.DateConflictOfInterestDeclarationChecked,
                 ResponsibleBodyResponseToTheConflictOfInterestRequestReceivedDate = supportProjectDto.ResponsibleBodyResponseToTheConflictOfInterestRequestReceivedDate,
-                HasShareEmailTemplateWithAdviser = supportProjectDto.HasShareEmailTemplateWithAdviser,
                 RemindAdviserToCopyRiseTeamWhenSentEmail = supportProjectDto.RemindAdviserToCopyRiseTeamWhenSentEmail,
                 IntroductoryEmailSentDate = supportProjectDto.IntroductoryEmailSentDate,
                 AdviserEmailAddress = supportProjectDto.AdviserEmailAddress,
@@ -218,6 +218,7 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Models.SupportProject
                 DateSupportOrganisationChosen = supportProjectDto.DateSupportOrganisationChosen,
                 SupportOrganisationName = supportProjectDto.SupportOrganisationName,
                 SupportOrganisationIdNumber = supportProjectDto.SupportOrganisationIdNumber,
+                SupportingOrganisationAddress = supportProjectDto.SupportingOrganisationAddress,
                 RegionalDirectorDecisionDate = supportProjectDto.RegionalDirectorDecisionDate,
                 HasSchoolMatchedWithSupportingOrganisation = supportProjectDto.HasSchoolMatchedWithSupportingOrganisation,
                 NotMatchingSchoolWithSupportingOrgNotes = supportProjectDto.NotMatchingSchoolWithSupportingOrgNotes,
@@ -234,6 +235,8 @@ namespace Dfe.ManageSchoolImprovement.Frontend.Models.SupportProject
                 DateSupportingOrganisationContactDetailsAdded = supportProjectDto.DateSupportingOrganisationContactDetailsAdded,
                 SupportingOrganisationContactName = supportProjectDto.SupportingOrganisationContactName,
                 SupportingOrganisationContactEmailAddress = supportProjectDto.SupportingOrganisationContactEmailAddress,
+                SupportingOrganisationContactPhone = supportProjectDto.SupportingOrganisationContactPhone,
+                SupportingOrganisationContactAddress = supportProjectDto.SupportingOrganisationAddress,
                 RegionalDirectorImprovementPlanDecisionDate = supportProjectDto.RegionalDirectorImprovementPlanDecisionDate,
                 HasApprovedImprovementPlanDecision = supportProjectDto.HasApprovedImprovementPlanDecision,
                 HasSavedImprovementPlanInSharePoint = supportProjectDto.HasSavedImprovementPlanInSharePoint,