+++ src/Dfe.ManageSchoolImprovement/Pages/TaskList/ChoosePreferredSupportingOrganisation/EnterSupportingOrganisationLocalAuthorityDetails.cshtml.cs
@@ -92,7 +92,13 @@ public class EnterSupportingOrganisationLocalAuthorityDetailsModel(
             LaCode,
             SupportProject?.SupportOrganisationType, // OrganisationType is maintained from the previous page
             DateSupportOrganisationConfirmed,
-            SupportProject?.AssessmentToolTwoCompleted);
+            SupportProject?.AssessmentToolTwoCompleted,
+            SupportProject?.SupportingOrganisationAddress,
+            SupportProject?.SupportingOrganisationContactName,
+            SupportProject?.SupportingOrganisationContactEmailAddress,
+            SupportProject?.SupportingOrganisationContactPhone,
+            SupportProject?.SupportingOrganisationAddress,
+            SupportProject?.DateSupportingOrganisationContactDetailsAdded);
 
         var result = await mediator.Send(command, cancellationToken);